requirements.txt
pandas
sqlalchemy
pyodbc
python-dotenv
snowflake-connector-python[pandas]